+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Refactored
 
-- Refactored the tip box for improved consistency and maintainability.
+- Refactored the tip box and image gallery for improved consistency and maintainability.
 
 ### Added